HelloWorld: previous deploy failed, mismatching script file names
The IntegrationFlow was referring to `thescript.groovy` while the
script was actually called `dosomething.groovy`.

SAP correctly identified the IntegrationFlow was invalid, and failed the
deployment.

Getting this kind of warning at Upload time might be nice. Maybe we can
do something with the "Problems" tab, that might have an API?